+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added support for socket asynchronous API for `recv` and `recvfrom`.
 - Added `net:gethostname/0` on platforms with gethostname(3).
 - Added `externalterm_to_term_with_roots` to efficiently preserve roots when allocating memory for external terms.
+- Added `atomvm:get_creation/0`, equivalent to `erts_internal:get_creation/0`
 
 ### Changed